Refactor all element specific logic out to a set of traits
