+++ src/Context/Awareness/CacheAwareContextTrait.php
@@ -1,0 +1,12 @@
+<?php
+namespace PaulGibbs\WordpressBehatExtension\Context\Awareness;
+
+trait CacheAwareContextTrait {
+    /**
+     * Clear object cache.
+     */
+    public function clearCache()
+    {
+        $this->getDriver()->cache->clear();
+    }
+}
