+++ src/Context/Awareness/ThemeAwareContextTrait.php
@@ -1,0 +1,20 @@
+<?php
+namespace PaulGibbs\WordpressBehatExtension\Context\Awareness;
+
+/**
+ *
+ * @author voddenr
+ *        
+ */
+trait ThemeAwareContextTrait {
+    /**
+     * Switch active theme.
+     *
+     * @param string $theme
+     */
+    public function switchTheme($theme)
+    {
+        $this->getDriver()->theme->change($theme);
+    }
+}
+
